+++ src/NaadSocketClient.php
@@ -132,14 +132,26 @@ class NaadSocketClient
      */
     protected function insertAlert( SimpleXMLElement $xml )
     {
+        $alert = null;
+
         try {
             $alert = Alert::fromXml($xml);
+        } catch(Exception $e) {
+            $alertId = $alert ? $alert->getId() : 'unknown';
+            $this->logger->critical($e->getMessage());
+            $this->logger->critical(
+                'Could not parse alert XML.'
+            );
+            throw $e;
+        }
+        try {
             $this->database->insertAlert($alert);
         } catch ( Exception $e ) {
+            $alertId = $alert ? $alert->getId() : 'unknown';
             $this->logger->critical($e->getMessage());
             $this->logger->critical(
                 'Could not connect to database or insert Alert ({id}).',
-                [ 'id' => $alert->getId() ]
+                [ 'id' => $alertId ]
             );
             throw $e;
         }